requirements.txt
docopt
pyyaml
attrdict
inquirer
